+++ src/main/kotlin/status-bar.kt
@@ -1,5 +1,0 @@
-// package com.github.clojure-lsp.intellij.extension.status-bar-widget-state
-
-// class ClojureLSPStatusBarWidgetState : WidgetState {
-
-// }