+++ src/main/resources/project-templates/clojure/src/project_name/main.clj
@@ -1,0 +1,5 @@
+(ns project-name.main)
+
+(defn -main [& args]
+  (println args)
+  0)