+++ src/main/resources/project-templates/clojure/test/project_name/main_test.clj
@@ -1,0 +1,7 @@
+(ns project-name.main-test
+  (:require
+   [clojure.test :refer [deftest is]]
+   [project-name.main :as main]))
+
+(deftest main-test
+  (is (= 0 (main/-main))))