+++ src/main/resources/project-templates/lein/test/project_name/core_test.clj
@@ -1,0 +1,7 @@
+(ns project-name.core-test
+  (:require [clojure.test :refer [deftest is testing]]
+            [project-name.core :as core]))
+
+(deftest a-test
+  (testing "FIXME, I fail."
+    (is (= 0 (core/foo 1)))))